requirements.txt
pandas
numpy<2.0
openpyxl
chromadb>=0.4.22,<0.5.0
google-generativeai>=0.3.0,<0.4.0
python-dotenv>=1.0.0,<2.0.0
fastapi>=0.100.0,<0.112.0
uvicorn[standard]>=0.20.0,<0.30.0
pydantic>=2.0.0,<3.0.0